Refactor Parser class
