+++ src/main/java/alice/task/Event.java
@@ -6,6 +6,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+import alice.parser.DateParser;
+import alice.parser.TaskParser;
+
 /** Task with /from and /to fields. */
 public class Event extends Task {
     private LocalDateTime from;
@@ -14,7 +17,7 @@ public class Event extends Task {
     public Event(String line) throws InvalidTaskException {
         super(line);
 
-        Map<String, String> flags = Parser.parseFlags(line);
+        Map<String, String> flags = TaskParser.parseFlags(line);
         if (!flags.containsKey("from")) {
             throw new InvalidTaskException("Missing /from flag.");
         }
@@ -23,12 +26,12 @@ public class Event extends Task {
         }
 
         try {
-            this.from = Parser.parseDateString(flags.get("from"));
+            this.from = DateParser.parseDateString(flags.get("from"));
         } catch (DateTimeParseException exception) {
             throw new InvalidTaskException("Invalid /from datetime. Should be yyyy-mm-dd hh:mm`");
         }
         try {
-            this.to = Parser.parseDateString(flags.get("to"));
+            this.to = DateParser.parseDateString(flags.get("to"));
         } catch (DateTimeParseException exception) {
             throw new InvalidTaskException("Invalid /to datetime. Should be yyyy-mm-dd hh:mm`");
         }
@@ -45,13 +48,13 @@ public class Event extends Task {
         keyValuePairs.add("\"taskType\": \"event\"");
         keyValuePairs.add(String.format("\"description\": \"%s\"", description));
         keyValuePairs.add(String.format("\"isCompleted\": \"%s\"", isCompleted));
-        keyValuePairs.add(String.format("\"from\": \"%s\"", Parser.toDateString(from)));
-        keyValuePairs.add(String.format("\"to\": \"%s\"", Parser.toDateString(from)));
+        keyValuePairs.add(String.format("\"from\": \"%s\"", DateParser.toDateString(from)));
+        keyValuePairs.add(String.format("\"to\": \"%s\"", DateParser.toDateString(from)));
         return String.format("{%s}", String.join(", ", keyValuePairs));
     }
 
     public static Task fromJsonString(String jsonString) throws InvalidTaskException {
-        Map<String, String> arguments = Parser.parseJsonString(jsonString);
+        Map<String, String> arguments = TaskParser.parseJsonString(jsonString);
         String inputLine = String.format(
             "event %s /from %s /to %s",
             arguments.get("description"),
